Rename `ThreadModel` to `Threading` and nest `TabCompletionsInstallation`
